chore: minor reorganization
