+++ packages/blockly/scripts/lib/package_json.mjs
@@ -14,10 +14,6 @@ import {fromRoot} from './fs_utils.mjs';
 /**
  * Load and return the contents of package.json.
  *
- * The file is re-read on every call so that callers always see
- * up-to-date data, even if it has been modified (e.g. to bump the
- * version number) since the script started.
- *
  * @returns {object} The parsed contents of package.json.
  */
 export function getPackageJson() {
